Moved EbolaRiskMapContainer styled-component to EbolaRiskMap/styles.js
