+++ src/components/EbolaRiskMap/__tests__/EbolaRiskMapContainer.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { shallow } from "enzyme";
 import { ThemeProvider as StyledComponentsProvider } from "styled-components";
 import styledComponentsTheme from "../../../assets/sc-theme";
-import { EbolaRiskMapContainer } from "../MapContainers";
+import { EbolaRiskMapContainer } from "../styles";
 
 test("render EbolaRiskMapContainer", () => {
   shallow(
